requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.1
Django==1.11
django-cors-headers==2.1.0
djangorestframework==3.6.3
Greeting==1.0.1
gunicorn==19.6.0
httpie==0.9.9
idna==2.6
nltk==3.2.4
numpy==1.13.1
oauthlib==2.0.2
psycopg2==2.6.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
twython==3.5.0
urllib3==1.22
whitenoise==3.3.0
